# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kimai - API",
        "description": "JSON API for the Kimai time-tracking software. Read our [API documentation](https://www.kimai.org/documentation/rest-api.html) and download the [Open API definition](doc.json) to import into your API client.\n",
        "version": "1.0"
    },
    "paths": {
...
        "/api/actions/project/{id}/{view}/{locale}": {
            "get": {
                "tags": [
                    "Actions"
                ],
                "summary": "Get all item actions for the given Project [for internal use]",
                "operationId": "get_get_project_actions",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project ID to fetch",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "\\d+"
                        }
                    },
                    {
                        "name": "view",
                        "in": "path",
                        "description": "View to display the actions at (e.g. index, custom)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Language to translate the action title to (e.g. de, en)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns item actions for the project",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageAction"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/projects": {
            "get": {
                "tags": [
                    "Project"
                ],
                "summary": "Returns a collection of projects (which are visible to the user)",
                "operationId": "get_get_projects",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "query",
                        "description": "Customer ID to filter projects",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "string",
                            "default": null,
                            "pattern": "\\d+"
                        }
                    },
                    {
                        "name": "customers[]",
                        "in": "query",
                        "description": "List of customer IDs to filter, e.g.: customers[]=1&customers[]=2",
                        "required": false,
                        "allowEmptyValue": true,
                        "explode": true,
                        "schema": {
                            "type": "array",
                            "items": {},
                            "default": [],
                            "pattern": "\\d+"
                        }
                    },
                    {
                        "name": "visible",
                        "in": "query",
                        "description": "Visibility status to filter projects: 1=visible, 2=hidden, 3=both",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "string",
                            "default": 1,
                            "pattern": "1|2|3"
                        }
                    },
                    {
                        "name": "start",
                        "in": "query",
                        "description": "Only projects that started before this date will be included. Allowed format: HTML5 (default: now, if end is also empty)",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "string",
                            "default": null
                        }
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "description": "Only projects that ended after this date will be included. Allowed format: HTML5 (default: now, if start is also empty)",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "string",
                            "default": null
                        }
                    },
                    {
                        "name": "ignoreDates",
                        "in": "query",
                        "description": "If set, start and end are completely ignored. Allowed values: 1 (default: off)",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "string",
                            "default": null
                        }
                    },
                    {
                        "name": "globalActivities",
                        "in": "query",
                        "description": "If given, filters projects by their 'global activity' support. Allowed values: 1 (supports global activities) and 0 (without global activities) (default: all)",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "string",
                            "default": null,
                            "pattern": "0|1"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "The result order. Allowed values: ASC, DESC (default: ASC)",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "string",
                            "default": null,
                            "pattern": "ASC|DESC"
                        }
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "The field by which results will be ordered. Allowed values: id, name, customer (default: name)",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "string",
                            "default": null,
                            "pattern": "id|name|customer"
                        }
                    },
                    {
                        "name": "term",
                        "in": "query",
                        "description": "Free search term",
                        "required": false,
                        "allowEmptyValue": false,
                        "schema": {
                            "type": "string",
                            "default": null
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a collection of projects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProjectCollection"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Project"
                ],
                "summary": "Creates a new project",
                "description": "Creates a new project and returns it afterwards",
                "operationId": "post_post_project",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectEditForm"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns the new created project",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectEntity"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/projects/{id}": {
            "get": {
                "tags": [
                    "Project"
                ],
                "summary": "Returns one project",
                "operationId": "get_get_project",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns one project entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectEntity"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Project"
                ],
                "summary": "Delete an existing project",
                "description": "[DANGER] This will also delete ALL linked activities and timesheets.\nMaybe use `PATCH` instead and mark it as inactive with `visible=false`?",
                "operationId": "delete_delete_project",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project ID to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete one project"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Project"
                ],
                "summary": "Update an existing project",
                "description": "Update an existing project, you can pass all or just a subset of all attributes",
                "operationId": "patch_patch_project",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project ID to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectEditForm"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns the updated project",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectEntity"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/teams/{id}/projects/{projectId}": {
            "post": {
                "tags": [
                    "Team"
                ],
                "summary": "Grant the team access to a project",
                "operationId": "post_post_team_project",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The team that is granted access",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "\\d+"
                        }
                    },
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "The project to grant acecess to (Project ID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Adds a new project to a team.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Team"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Team"
                ],
                "summary": "Revokes access for a project from a team",
                "operationId": "delete_delete_team_project",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The team whose permission will be revoked",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "\\d+"
                        }
                    },
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "The project to remove (Project ID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Removes a project from the team.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Team"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TeamMember": {
                "required": [
                    "user"
                ],
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "teamlead": {
                        "type": "boolean",
                        "default": false
                    }
                },
                "type": "object"
            },
...
            "Team": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "title": "Team name",
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 2
                    },
                    "members": {
                        "title": "All team member (including team leads)",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TeamMember"
                        },
                        "minItems": 1
                    },
                    "customers": {
                        "title": "Customers assigned to the team",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Customer"
                        }
                    },
                    "projects": {
                        "title": "Projects assigned to the team",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Project"
                        }
                    },
                    "activities": {
                        "title": "Activities assigned to the team",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    },
                    "color": {
                        "title": "The assigned color in HTML hex format, e.g. #dd1d00",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ProjectCollection": {
                "required": [
                    "name"
                ],
                "properties": {
                    "parentTitle": {
                        "type": "string"
                    },
                    "customer": {
                        "type": "integer"
                    },
                    "id": {
                        "title": "Unique Project ID",
                        "type": "integer"
                    },
                    "name": {
                        "title": "Project name",
                        "type": "string",
                        "maxLength": 150,
                        "minLength": 2
                    },
                    "start": {
                        "title": "Project start date (times before this date cannot be recorded)",
                        "description": "Attention: Accessor MUST be used, otherwise date will be serialized in UTC.",
                        "type": "string",
                        "format": "date-time"
                    },
                    "end": {
                        "title": "Project end time (times after this date cannot be recorded)",
                        "description": "Attention: Accessor MUST be used, otherwise date will be serialized in UTC.",
                        "type": "string",
                        "format": "date-time"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "visible": {
                        "title": "If the project is not visible, times cannot be recorded",
                        "type": "boolean",
                        "default": true
                    },
                    "billable": {
                        "type": "boolean",
                        "default": true
                    },
                    "metaFields": {
                        "title": "Meta fields registered with the project",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProjectMeta"
                        }
                    },
                    "teams": {
                        "title": "Teams with access to the project",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Team"
                        }
                    },
                    "globalActivities": {
                        "title": "Whether this project allows booking of global activities",
                        "type": "boolean",
                        "default": true
                    },
                    "number": {
                        "type": "string",
                        "maxLength": 10
                    },
                    "color": {
                        "title": "The assigned color in HTML hex format, e.g. #dd1d00",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ProjectEntity": {
                "required": [
                    "name"
                ],
                "properties": {
                    "parentTitle": {
                        "type": "string"
                    },
                    "customer": {
                        "type": "integer"
                    },
                    "id": {
                        "title": "Unique Project ID",
                        "type": "integer"
                    },
                    "name": {
                        "title": "Project name",
                        "type": "string",
                        "maxLength": 150,
                        "minLength": 2
                    },
                    "orderNumber": {
                        "title": "Project order number",
                        "type": "string",
                        "maxLength": 50
                    },
                    "orderDate": {
                        "title": "Order date for the project",
                        "description": "Attention: Accessor MUST be used, otherwise date will be serialized in UTC.",
                        "type": "string",
                        "format": "date-time"
                    },
                    "start": {
                        "title": "Project start date (times before this date cannot be recorded)",
                        "description": "Attention: Accessor MUST be used, otherwise date will be serialized in UTC.",
                        "type": "string",
                        "format": "date-time"
                    },
                    "end": {
                        "title": "Project end time (times after this date cannot be recorded)",
                        "description": "Attention: Accessor MUST be used, otherwise date will be serialized in UTC.",
                        "type": "string",
                        "format": "date-time"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "visible": {
                        "title": "If the project is not visible, times cannot be recorded",
                        "type": "boolean",
                        "default": true
                    },
                    "billable": {
                        "type": "boolean",
                        "default": true
                    },
                    "metaFields": {
                        "title": "Meta fields registered with the project",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProjectMeta"
                        }
                    },
                    "teams": {
                        "title": "Teams with access to the project",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Team"
                        }
                    },
                    "globalActivities": {
                        "title": "Whether this project allows booking of global activities",
                        "type": "boolean",
                        "default": true
                    },
                    "number": {
                        "type": "string",
                        "maxLength": 10
                    },
                    "budget": {
                        "title": "The total monetary budget, will be zero if not configured.",
                        "type": "number",
                        "format": "float",
                        "default": 0
                    },
                    "timeBudget": {
                        "title": "The time budget in seconds, will be zero if not configured.",
                        "type": "integer",
                        "default": 0,
                        "maximum": 2145600000,
                        "minimum": 0
                    },
                    "budgetType": {
                        "title": "The type of budget:\n - null      = default / full time\n - month     = monthly budget",
                        "type": "string"
                    },
                    "color": {
                        "title": "The assigned color in HTML hex format, e.g. #dd1d00",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ProjectEditForm": {
                "required": [
                    "name",
                    "customer"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "number": {
                        "type": "string"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "invoiceText": {
                        "type": "string"
                    },
                    "orderNumber": {
                        "type": "string"
                    },
                    "orderDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "start": {
                        "type": "string",
                        "format": "date"
                    },
                    "end": {
                        "type": "string",
                        "format": "date"
                    },
                    "customer": {
                        "description": "Customer ID",
                        "type": "integer"
                    },
                    "color": {
                        "description": "The hexadecimal color code (default: auto-calculated by name)",
                        "type": "string"
                    },
                    "globalActivities": {
                        "type": "boolean"
                    },
                    "visible": {
                        "type": "boolean"
                    },
                    "billable": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "Customer": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 150,
                        "minLength": 2
                    },
                    "number": {
                        "type": "string",
                        "maxLength": 50
                    },
                    "comment": {
                        "type": "string"
                    },
                    "visible": {
                        "type": "boolean",
                        "default": true
                    },
                    "billable": {
                        "type": "boolean",
                        "default": true
                    },
                    "color": {
                        "title": "The assigned color in HTML hex format, e.g. #dd1d00",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "PageAction": {
                "properties": {
                    "id": {
                        "title": "ID of the action",
                        "type": "string"
                    },
                    "title": {
                        "title": "Translated title to show the user",
                        "type": "string"
                    },
                    "url": {
                        "title": "URL of the action",
                        "type": "string"
                    },
                    "class": {
                        "title": "HTML classes to be used",
                        "type": "string"
                    },
                    "attr": {
                        "title": "HTML (data) attributes to render the action",
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    },
                    "divider": {
                        "title": "Whether to render a divider before this item",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "ProjectMeta": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "Name of the meta (custom) field",
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 2
                    },
                    "value": {
                        "title": "Value of the meta (custom) field",
                        "type": "string",
                        "maxLength": 65535
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "API Token",
                "bearerFormat": "KIMAI",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearer": []
        }
    ],
...
}